Fix includes

+++ test/index.ts
@@ -1,7 +1,7 @@
 import fs from "node:fs"
 import path from "node:path"
 
-import staticbuild from "../src/staticbuild"
+import staticbuild, { StaticBuildOptions } from "../src/staticbuild"
 import { scan } from "../src/fs"
 
 function clamp(value: number, min: number, max: number) {
@@ -51,6 +51,7 @@ function expectFilesEqual(actualFilePath: string, expectedFilePath: string) {
   if (isDifferent) {
     console.log(`\x1b[38;2;0;255;0mExpected:\n${expectedContents}\x1b[0m`)
     console.log(`\x1b[38;2;255;0;0mActual:\n${actualContents}\x1b[0m`)
+    throw Error("Files are different")
   }
 }
 
@@ -79,20 +80,33 @@ function expectDirectoriesEqual(actualDirectoryPath: string, expectedDirectoryPa
 }
 
 async function test() {
+  // Remove the first 2 arguments that nodejs provides.
+  const args = process.argv.splice(2, process.argv.length)
+
+  const logger: StaticBuildOptions["logger"] = {
+    info: (...messages: unknown[]) => {},
+    warn: console.log,
+    error: console.log,
+    time: (name: string) => {},
+    timeEnd: (name: string) => {},
+  }
+
   for (const directory of scan("./test", [], { recursive: false })) {
+    if (args[0] && args[0] !== directory.name) continue
+
     if (!directory.isDirectory) continue
     if (directory.name.startsWith("x_")) continue
 
     console.log(`Test: ${directory.name}`)
 
-    const inputDirectory = path.join(directory.path, "input")
-    const outputDirectory = path.join(directory.path, "output")
+    const inputDirectory = path.join(process.cwd(), directory.path, "input")
+    const outputDirectory = path.join(process.cwd(), directory.path, "output")
 
     if (fs.existsSync(outputDirectory)) {
       fs.rmSync(outputDirectory, { recursive: true })
     }
-    
-    await staticbuild({ inputDirectory, outputDirectory })
+
+    await staticbuild({ inputDirectory, outputDirectory, logger })
 
     const expectedDirectory = path.join(directory.path, "expected")
     expectDirectoriesEqual(outputDirectory, expectedDirectory)
